moved ControlInv
