+++ mtpy/modeling/ModEM/__init__.py
@@ -3,5 +3,6 @@ from .station import Stations
 from .data import Data
 from .model import Model
 from .residual import Residual
+from .controlinv import ControlInv
 
-__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual']
+__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual', 'ControlInv']
